# Faster ParameterMap