--- a/Framework/Geometry/src/Instrument/ParameterMap.cpp
+++ b/Framework/Geometry/src/Instrument/ParameterMap.cpp
@@ -348,7 +348,11 @@ void ParameterMap::add(const IComponent *comp,
   if (existing_par != m_map.end()) {
     existing_par->second = par;
   } else {
+#if TBB_VERSION_MAJOR >= 4 && TBB_VERSION_MINOR >= 4
+    m_map.emplace(comp->getComponentID(), par);
+#else
     m_map.insert(std::make_pair(comp->getComponentID(), par));
+#endif
   }
 }
 
@@ -1022,7 +1026,11 @@ void ParameterMap::copyFromParameterMap(const IComponent *oldComp,
   for (const auto &oldParameterName : oldParameterNames) {
     Parameter_sptr thisParameter = oldPMap->get(oldComp, oldParameterName);
     // Insert the fetched parameter in the m_map
+#if TBB_VERSION_MAJOR >= 4 && TBB_VERSION_MINOR >= 4
+    m_map.emplace(newComp->getComponentID(), thisParameter);
+#else
     m_map.insert(std::make_pair(newComp->getComponentID(), thisParameter));
+#endif
   }
 }