… CMakeLists.txt
… Framework/Geometry/CMakeLists.txt
… Framework/Kernel/CMakeLists.txt
… Framework/Kernel/inc/MantidKernel/Unit.h
… Framework/Kernel/src/Unit.cpp
… Framework/Kernel/test/CMakeLists.txt
… Framework/PythonInterface/mantid/geometry/CMakeLists.txt
… Framework/PythonInterface/mantid/kernel/CMakeLists.txt
… buildconfig/CMake/Bootstrap.cmake
… buildconfig/CMake/FindTBB.cmake
… buildconfig/CMake/WindowsNSIS.cmake