Description of work.

This improves the performance of SCDCalibratePanels and other algorithms which access the ParameterMap from multiple threads. The `tbb::concurrent_unordered_multimap` supports concurrent insertion and traversal, while critical sections enclose all calls to `unsafe_erase(...)`. FindTBB.cmake was copied from VTK.

`boost::iequal`, which uses [locales and contains a mutex](http://stackoverflow.com/questions/1771386/explain-the-need-for-mutexes-in-locales-please) was replaced with `strcasecmp` on Linux and `_stricmp` on Windows.

The critical section in SCDCalibratePanels was eliminated by moving `push_back` outside of the parallel loop and accessing the vector by its index.

Fixed a bug in `ParameterMap::clearParametersByName` where not all names associated with a particular component were being checked/cleared. 

![scdcalibratepanels](https://cloud.githubusercontent.com/assets/5920552/18166069/ecccd3c2-7016-11e6-96b1-4472bd1a9670.png)

SHA=5135bc0
1 3m10.239s
2 2m50.114s
4 2m48.759s
8 2m48.095s
16 2m50.160s
32 2m58.167s
64 3m2.704s
(unspecified) 3m5.546s

mantidpythonnightly August 30, 2016
1 1m46.937s 1m46.341s
2 1m25.669s 1m23.106s
4 1m44.806s 1m43.665s
8 1m44.524s 1m42.377s
16 1m52.687s 1m48.501s
32 1m58.103s 1m56.550s
64 2m0.437s 2m0.435s
(unspecified) 1m58.707s 1m58.828s

this branch
1 1m16.127s 1m16.631s
2  59.282s 59.166s
4  57.371s 57.686s
8   50.838s 50.196s
16 48.046s 48.653s 
32 47.882s 46.000s
64 43.811s 43.856s
(unspecified) 44.636s 45.408s

**To test:**

<!-- Instructions for testing. -->

Fixes #17308.

<!-- RELEASE NOTES
Either edit the file in docs/source/release/... and it will be in your pull request or state
*Does not need to be in the release notes.*
-->

---

#### Reviewer ####

Please comment on the following ([full description](http://www.mantidproject.org/Individual_Ticket_Testing)):

##### Code Review #####

- [ ] Is the code of an acceptable quality?
- [ ] Does the code conform to the [coding standards](http://www.mantidproject.org/Coding_Standards)? Is it well structured with small focussed classes/methods/functions?
- [ ] Are there unit/system tests in place? Are the unit tests small and test the a class in isolation?
- [ ] If there are changes in the release notes then do they describe the changes appropriately?

##### Functional Tests #####

- [ ] Do changes function as described? Add comments below that describe the tests performed?
- [ ] How do the changes handle unexpected situations, e.g. bad input?
- [ ] Has the relevant documentation been added/updated?
- [ ] Is user-facing documentation written in a user-friendly manner?
- [ ] Has developer documentation been updated if required?

- Does everything look good? Comment with the ship it emoji but don't merge. A member of `@mantidproject/gatekeepers` will take care of it.