@quantumsteve 2016-08-26
I successfully built this branch on Windows by copying headers and libraries from the [latest stable release of tbb](https://www.threadingbuildingblocks.org/sites/default/files/software_releases/windows/tbb44_20160526oss_win_0.zip) into the appropriate directories in `ThirdParty`. So far I've been unable to push that branch to GitHub.