@VickieLynch 2016-09-01
From 1min30sec to 36 sec with same answers for calibration.  Great work Steven :shipit: